fix moved to common file.
